Reorganize and modernize aliases
`:deps` -> `:add-lib` so it no longer interferes with the installed 
`:deps` alias in the Clojure CLI. Also made it clear this is 
experimental and may break or go away.

`:rebl` and `:rebl-win` are now the only REBL aliases because as of 
0.9.226 it is packaged to work with `java8` (with bundled OpenJFX) and 
`openjfx15ea` without needing separate dependencies.

Removed `:proto` as ProtoREPL is unmaintained.

Ensure the README matches the actual aliases provided and organize them 
in likely order of use/importance.